👀 [Feat] 시력검사 테스트 시작화면으로 이동 구현
